==> code/1_Data_Kekerasan_Anak_Clean.ipynb <==
#Import library yang dibutuhkan
import pandas as pd
import numpy as np

#Buka data pengaduan anak di pandas
data_pengaduan_anak = pd.read_excel('Data Pengaduan Anak.xlsx')
data_pengaduan_anak.head()

#Melt data agar data bisa lebih mudah dianalisis
data_pengaduan_anak_melt = (pd.melt(data_pengaduan_anak, id_vars=['No','KASUS PERLINDUNGAN ANAK'], var_name='Tahun', value_name='total'))
data_pengaduan_anak_melt.head()

#hilangkan data data null
data_pengaduan_anak_melt = data_pengaduan_anak_melt[data_pengaduan_anak_melt['No'].notna()]

#Hilangkan data data "total sub group data"
data_pengaduan_anak_melt.drop(data_pengaduan_anak_melt.loc[data_pengaduan_anak_melt['No']==1].index, inplace=True)
data_pengaduan_anak_melt.drop(data_pengaduan_anak_melt.loc[data_pengaduan_anak_melt['No']==2].index, inplace=True)
data_pengaduan_anak_melt.drop(data_pengaduan_anak_melt.loc[data_pengaduan_anak_melt['No']==3].index, inplace=True)
data_pengaduan_anak_melt.drop(data_pengaduan_anak_melt.loc[data_pengaduan_anak_melt['No']==4].index, inplace=True)
data_pengaduan_anak_melt.drop(data_pengaduan_anak_melt.loc[data_pengaduan_anak_melt['No']==5].index, inplace=True)
data_pengaduan_anak_melt.drop(data_pengaduan_anak_melt.loc[data_pengaduan_anak_melt['No']==6].index, inplace=True)
data_pengaduan_anak_melt.drop(data_pengaduan_anak_melt.loc[data_pengaduan_anak_melt['No']==7].index, inplace=True)
data_pengaduan_anak_melt.drop(data_pengaduan_anak_melt.loc[data_pengaduan_anak_melt['No']==8].index, inplace=True)
data_pengaduan_anak_melt.drop(data_pengaduan_anak_melt.loc[data_pengaduan_anak_melt['No']==9].index, inplace=True)
data_pengaduan_anak_melt.drop(data_pengaduan_anak_melt.loc[data_pengaduan_anak_melt['No']==10].index, inplace=True)

# Buat fungsi untuk mencari data jawa barat saja dengan asumsi jumlah penduduk jawa barat / total nasional
def data_jawa_barat(value):
    return round(value *(49565/269603))

# buat table baru untuk menghitung total jawa barat saja
data_pengaduan_anak_melt['total_jawa_barat'] = data_pengaduan_anak_melt['total'].apply(data_jawa_barat)


#Buka data pengaduan anak di pandas untuk mendapatkan granularity dan sebaran data anak di jawa barat
data_ibu_hamil = pd.read_csv('jumlah_ibu_hamil_berdasarkan_kabupatenkota_data.csv')
data_ibu_hamil.head()

#drop data yang tidak perlu
data_ibu_hamil_percentage = data_ibu_hamil.drop(columns=['id','kode_provinsi','nama_provinsi','nama_kabupaten_kota','satuan'])
data_ibu_hamil_percentage

#Groupby data untuk mendapatkan hasil pengelompokan berdasarkan tahun dan kota
data_ibu_hamil_percentage = data_ibu_hamil_percentage.groupby(['tahun','kode_kabupaten_kota'],).sum()
data_ibu_hamil_percentage

# cari persentase tiap tahun dan kota untuk setiap anak
data_ibu_hamil_percentage = data_ibu_hamil_percentage.div(data_ibu_hamil_percentage.sum(level=0), level=0)
data_ibu_hamil_percentage

#Simpan data ibu Hamil ke dalam csv untuk backup
data_ibu_hamil_percentage.to_csv('data_ibu_hamil_percentage.csv',index=True)


# input data ibu hamil jabar
data_ibu_hamil_jabar = pd.read_csv('data_ibu_hamil_percentage.csv')
data_ibu_hamil_jabar.head()

#merge data pengaduan anak jabar dengan data ibu hamil untuk mendapatkan perkiraan data pengaduan anak per kota kabupaten
data_pengaduan_anak_jabar = pd.merge(data_pengaduan_anak_melt, data_ibu_hamil_jabar, how='left', left_on='Tahun', right_on= 'tahun')

# kali jumlah data pengaduan anak jawa barat dengan persentase data anak tiap kabupaten dan tiap tahun
data_pengaduan_anak_jabar ['total_kota_jabar'] = data_pengaduan_anak_jabar ['total_jawa_barat'] * data_pengaduan_anak_jabar ['jumlah_ibu_hamil']

# drop data data yang tidak dibutuhkan
data_pengaduan_anak_jabar_clean = data_pengaduan_anak_jabar.drop(columns=['tahun','total','jumlah_ibu_hamil'])
data_pengaduan_anak_jabar_clean

# save data kedalam csv untuk di proses lebih lanjut
data_pengaduan_anak_jabar_clean.to_csv('data_pengaduan_anak_jabar_clean.csv',index=False)
