==> code/2_Data_Kekerasan_Keluarga_Join.ipynb <==
#Import library yang dibutuhkan
import pandas as pd
import numpy as np

#Buka data pengaduan anak yang sudah kita cleaning di proses 1 di pandas
data_pengaduan_anak = pd.read_csv('data_pengaduan_anak_jabar_clean.csv')
data_pengaduan_anak.head()

#Buka data kekerasan rumah tangga di pandas
data_kdrt = pd.read_csv('jml_korban_kekerasan_dalam_rumah_tangga_bedasarkan_jk_data.csv')
data_kdrt.head()

# drop data yang tidak dibutuhkan di data kdrt
data_kdrt = data_kdrt.drop(columns=['id','kode_provinsi','nama_provinsi','nama_kabupaten_kota','satuan'])
data_kdrt

# manipulasi data kdrt dan disimulasikan data 2021 yang ada diubah menjadi data 2020 agar bisa berkorelasi dengan data kekerasan anak
data_kdrt['tahun'] = data_kdrt['tahun'].map({2018:2018 , 2021:2020})

# grouping data untuk mendapat data yang lebih bersih
data_kdrt_group = data_kdrt.groupby(['tahun','kode_kabupaten_kota']).sum()
data_kdrt_group.head()

# merge data antara data pengaduan anak dan data kdrt
data_kdrt_dan_pengaduan_anak = pd.merge(data_pengaduan_anak, data_kdrt_group, how='left', left_on=['Tahun','kode_kabupaten_kota'], right_on= ['tahun','kode_kabupaten_kota'])

# drop kolom yang tidak digunakan
data_kdrt_dan_pengaduan_anak = data_kdrt_dan_pengaduan_anak.drop(columns=['total_jawa_barat'])
data_kdrt_dan_pengaduan_anak

# Rename column name agar memudahkan disaat analisa data kedepannya
data_kdrt_dan_pengaduan_anak.rename(columns = {'total_kota_jabar':'total_kekerasan_anak', 'jumlah_korban':'total_kdrt'}, inplace = True)

# save data kedalam csv untuk analisa lebih lanjut
data_kdrt_dan_pengaduan_anak.to_csv('data_kdrt_dan_pengaduan_anak.csv',index=False)
