==> code/3_data_cleaning_index_kesehatan.ipynb <==
#Import library yang dibutuhkan
import pandas as pd
import numpy as np


# import data index kesehatan, dan jadikan sebagai data utama
data_index_kesehatan= pd.read_csv('indeks_kesehatan_berdasarkan_kabupatenkota_data.csv')
data_index_kesehatan.head()

# import data ims, jadikan data source kedua
data_ims= pd.read_csv('jml_kasus_penyakit_infeksi_menular_seksual_ims__kabupa_data.csv')
data_ims.head()

#drop column data yang tidak digunakan
data_ims = data_ims.drop(columns=['id','kode_provinsi','nama_provinsi','nama_kabupaten_kota','satuan'])
data_ims.head()

#merge data pertama, antara data index kesehatan dan data ims
data_join1 = pd.merge(data_index_kesehatan, data_ims, how='left', left_on=['tahun','kode_kabupaten_kota'], right_on= ['tahun','kode_kabupaten_kota'])
data_join1

# buka data balita gizi buruk 2019-2020
data_balita_gizi_buruk2019= pd.read_csv('jumlah_balita_berdasarkan_kategori_balita_gizi_buruk_data.csv')
data_balita_gizi_buruk2019.head()

#drop data 2020, dan hanya ambil data 2019 saja
data_balita_gizi_buruk2019.drop(data_balita_gizi_buruk2019.loc[data_balita_gizi_buruk2019['tahun']==2020].index, inplace=True)

#drop kolom yang tidak digunakan

data_balita_gizi_buruk2019 = data_balita_gizi_buruk2019.drop(columns=['id','kategori_gizi_buruk'])
data_balita_gizi_buruk2019.head()

#read data 2020-2021 dan drop kolom id
data_balita_gizi_buruk2020= pd.read_csv('jumlah_balita_bergizi_kurang_berdasarkan_kabupatenkota_data.csv')
data_balita_gizi_buruk2020 = data_balita_gizi_buruk2020.drop(columns=['id'])
data_balita_gizi_buruk2020.head()

# join data 2, antara data balita gizi buruk 2020 dan balita gizi buruk 2019
data_join2 = pd.concat([data_balita_gizi_buruk2020, data_balita_gizi_buruk2019], ignore_index = True)

# drop data yang tidak dibutuhkan
data_join2 = data_join2.drop(columns=['kode_provinsi','nama_provinsi','nama_kabupaten_kota','satuan'])
data_join2

# groupby data agar data lebih clean
data_join2 = data_join2.groupby(['tahun','kode_kabupaten_kota']).sum()
data_join2.head()

#data join 3, yaitu menggabungkan data join 1 (main) dan data join 2 (balita gizi buruk)
data_join3 = pd.merge(data_join1, data_join2, how='left', left_on=['tahun','kode_kabupaten_kota'], right_on= ['tahun','kode_kabupaten_kota'])
data_join3

# rename beberapa column untuk memudahkan analisa nantinya
data_join3.rename(columns = {'jumlah_kasus':'jumlah_kasus_ims', 'jumlah_balita':'jumlah_balita_giziburuk'}, inplace = True)

# masukkan data bayi gizi buruk
data_bayi_gizi_buruk= pd.read_csv('jumlah_bayi_bergizi_buruk_berdasarkan_kabupatenkota_data.csv')
data_bayi_gizi_buruk.head()

# drop column yang tidak dibutuhkan
data_bayi_gizi_buruk = data_bayi_gizi_buruk.drop(columns=['id','kode_provinsi','nama_provinsi','nama_kabupaten_kota','satuan'])

#data join 4, yaitu menggabungkan data join 3 (main) dan data join 4 (bayi gizi buruk)
data_join4 = pd.merge(data_join3, data_bayi_gizi_buruk, how='left', left_on=['tahun','kode_kabupaten_kota'], right_on= ['tahun','kode_kabupaten_kota'])
data_join4.head()

# masukkan data keluhan kesehatan
data_keluhan_kesehatan= pd.read_csv('persentase_penduduk_mempunyai_keluhan_kesehatan__kabup_data.csv')
data_keluhan_kesehatan.head()

# data drop untuk menghilangkan data data yang tidak digunakan
data_keluhan_kesehatan = data_keluhan_kesehatan.drop(columns=['id','kode_provinsi','nama_provinsi','nama_kabupaten_kota','satuan'])

#data join 5, yaitu menggabungkan data join 4 (main) dan data join 5 (keluhan kesehatan)
data_join5 = pd.merge(data_join4, data_keluhan_kesehatan, how='left', left_on=['tahun','kode_kabupaten_kota'], right_on= ['tahun','kode_kabupaten_kota'])
data_join5.head()

# rename data header untuk memudahkan analisa kedepannya
data_join5.rename(columns = {'jumlah_bayi':'jumlah_bayi_giziburuk', 'persentase_penduduk':'persentase_keluhan_kesehatan'}, inplace = True)

# simpan data dalam csv untuk analisa lebih lanjut kedepannya
data_join5.to_csv('data_index_kesehatan_dan_parameter.csv',index=False)
